Experimental DualShock 3 (wired) support
Based on libsickaxis 1.0 by Xerpi:
https://github.com/xerpi/libsicksaxis

"SickSaxis is a C library that lets you connect a Sixaxis or a DualShock3 to the Wii and read its data. It also lets you turn on and off the controller LEDs, enable or disable the rumble and get and set the controllers' bluetooth MAC address."

To enable the DS3 controller, press and hold both minus (-) and plus (+) buttons in sequence, then release. The DS3 led #1 should light up once, and then the four leds. After that, press the PS button, it should be enabled for in-game use by now. You can do this before loading any game, for example. As of now, it controls the 2-4 players at the same time.

Could and should be upgraded in the future to polish up this implementation, any help on this task will be welcome.