BS-X Satellaview engine upgrade
Satellaview emulation synced to latest Snes9x to date. This furthers up the game compatibility, by making use of the Satellaview BIOS. Though the file is optional, you'll want to use it for titles that make exclusive use of its functions. Some of the titles now playable are:

- BS Fire Emblem: Akaneia Saga (Episodes 1 to 4)
- Chou Aniki - Bakuretsu Rantou Hen
- Konae-chan no Dokidoki Penguin Kazoku Event-ban

And a lot of the digital magazines available that were distributed through the Satellite's service can be read in its original language now, as the BIOS embedded font is mostly used for these.

Check the instructions for its use in the README file.